requirements.txt
mysql-connector==2.2.9
phpserialize==1.3
itemadapter==0.1.0